+++ src/Telegram.Bot/Types/InlineQueryResults/InputContactMessageContent.cs
@@ -27,6 +27,12 @@ namespace Telegram.Bot.Types.InlineQueryResults
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
         public string LastName { get; set; }
 
+        /// <summary>
+        /// Optional. Additional data about the contact in the form of a vCard, 0-2048 bytes
+        /// </summary>
+        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
+        public string Vcard { get; set; }
+
         private InputContactMessageContent()
         { }