+++ test/Telegram.Bot.Tests.Integ/Sending Messages/SendingContactMessageTests.cs
@@ -42,9 +42,45 @@ namespace Telegram.Bot.Tests.Integ.Sending_Messages
             Assert.Equal(lastName, message.Contact.LastName);
         }
 
+        [OrderedFact(DisplayName = FactTitles.ShouldSendContactWithVCardd)]
+        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendContact)]
+        public async Task Should_Send_Contact_With_VCard()
+        {
+            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldSendContactWithVCardd);
+
+            const string vcard =
+                "BEGIN:VCARD" + "\n" +
+                "VERSION:2.1" + "\n" +
+                "N:Gump;Forrest;;Mr." + "\n" +
+                "FN:Forrest Gump" + "\n" +
+                "ORG:Bubba Gump Shrimp Co." + "\n" +
+                "TITLE:Shrimp Man" + "\n" +
+                "PHOTO;JPEG:https://upload.wikimedia.org/wikipedia/commons/9/95/TomHanksForrestGump94.jpg" + "\n" +
+                "TEL;WORK;VOICE:(111) 555-1212" + "\n" +
+                "TEL;HOME;VOICE:(404) 555-1212" + "\n" +
+                "ADR;HOME:;;42 Plantation St.;Baytown;LA;30314;United States of America" + "\n" +
+                "LABEL;HOME;ENCODING=QUOTED-PRINTABLE;CHARSET=UTF-8:42 Plantation St.=0D=0A=" + "\n" +
+                " Baytown, LA 30314=0D=0AUnited States of America" + "\n" +
+                "EMAIL:forrestgump@example.org" + "\n" +
+                "REV:20080424T195243Z" + "\n" +
+                "END:VCARD";
+
+            Message message = await BotClient.SendContactAsync(
+                /* chatId: */ _fixture.SupergroupChat,
+                /* phoneNumber: */ "+11115551212",
+                /* firstName: */ "Forrest",
+                vCard: vcard
+            );
+
+            Assert.Equal(MessageType.Contact, message.Type);
+            Assert.Equal(vcard, message.Contact.Vcard);
+        }
+
         private static class FactTitles
         {
             public const string ShouldSendContact = "Should send a contact";
+
+            public const string ShouldSendContactWithVCardd = "Should send a contact including his vCard";
         }
     }
 }