+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Parameter `vCard` to method `SendContactAsync`
+- Parameter `foursquareType` to method `SendVenueAsync`
+- Property `SendContactRequest.Vcard`
+- Property `SendVenueRequest.FoursquareType`
+- Property `InlineQueryResultContact.Vcard`
+- Property `InlineQueryResultVenue.FoursquareType`
+- Property `InputContactMessageContent.Vcard`
+- Property `InputVenueMessageContent.FoursquareType`
+- Property `Contact.Vcard`
+- Property `Venue.FoursquareType`
+- Enum value `MessageEntityType.Cashtag`
+
+### Changed
+
+- Marked `MessageType.Animation` as obsolete.
+
 ## [14.7.0] - 2018-07-29
 
 ### Added